chat

+++ mblogcommon/src/main/java/com/mtons/mblog/entity/GroupUser.java
@@ -11,5 +11,5 @@ public class GroupUser {
 
     private String userName;
 
-    private Integer chatGroupId;
+    private Long chatGroupId;
 }
